drf-lab1/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
django-heroku==0.0.0
djangorestframework==3.12.4
Faker==8.10.3
gunicorn==20.0.4
Markdown==3.3.4
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
validate-docbr==1.8.2
whitenoise==5.3.0
